requirements.txt
# Python dependencies for the FastAPI Allergy Scanner app
# Note: pyzbar requires the system library 'zbar' (e.g., on macOS: `brew install zbar`).

fastapi==0.111.0                 # Web framework for building the API (Pydantic v2-based)
uvicorn[standard]==0.30.6        # ASGI server with standard extras (uvloop, httptools, websockets, watchfiles)

sqlalchemy==2.0.32               # ORM and SQL toolkit (using async engine/session)
aiosqlite==0.20.0                # Async SQLite driver used by SQLAlchemy's aiosqlite dialect

python-jose[cryptography]==3.3.0 # JWT creation/verification using the 'cryptography' backend
passlib[bcrypt]==1.7.4           # Password hashing utilities with bcrypt algorithm

httpx==0.27.2                    # Async HTTP client for calling Open Food Facts API

pyzbar==0.1.9                    # Barcode decoding from images (wraps system 'zbar' library)
Pillow==10.4.0                   # Image processing (open/convert images for barcode decoding)

python-multipart==0.0.9          # Parse multipart/form-data (needed for file uploads in FastAPI)
python-dotenv==1.0.1             # Load environment variables from .env files

